requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
biwrap==0.1.6
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
exceptiongroup==1.2.0
fastapi
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grad-cam==1.4.8
grpcio==1.59.0
h5py==3.9.0
huggingface-hub
hydra-core==1.3.2
idna==3.4
imageio==2.31.6
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lightning
lightning-utilities
lime==0.2.0.1
llvmlite==0.41.1
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.0
pathtools==0.1.2
Pillow==10.0.1
protobuf==4.24.4
psutil==5.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic
pydantic_core
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.12.2
sentry-sdk==1.30.0
setproctitle==1.3.2
shap==0.43.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-plot==0.3.2
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.13.3
torch==2.0.0
torch-summary==1.4.5
torchaudio==2.0.0
torchmetrics==1.2.0
torchvision==0.15.0
tqdm==4.66.1
ttach==0.0.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.7
wandb==0.15.10
Werkzeug==3.0.0
wrapt==1.14.0
yarl==1.9.2
